tests: Separate mod
