+++ tests/part.rs
@@ -1,0 +1,176 @@
+// Copyright © 2025 The µcad authors <info@ucad.xyz>
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+#[cfg(test)]
+use microcad_lang::{eval::*, model_tree::*, resolve::*, src_ref::*, syntax::*, value::*};
+
+/// Helper function to create a qualified name from &str
+#[cfg(test)]
+fn qualified_name(s: &str) -> QualifiedName {
+    QualifiedName::no_ref(
+        s.split("::")
+            .map(|x| Identifier(Refer::none(x.into())))
+            .collect(),
+    )
+}
+
+/// Helper function to create a call argument list from &str
+#[cfg(test)]
+fn call_argument_value_list(s: &str, context: &mut Context) -> CallArgumentValueList {
+    use microcad_lang::parser::*;
+
+    let call_argument_list = Parser::parse_rule::<CallArgumentList>(Rule::call_argument_list, s, 0)
+        .expect("Valid CallArgumentList");
+
+    call_argument_list
+        .eval(context)
+        .expect("Valid call argument list")
+}
+
+#[test]
+fn part_implicit_init_call() {
+    let mut context = crate::context_for_file("syntax/part/implicit_init.µcad");
+
+    let node = context.lookup(&qualified_name("a")).expect("Node expected");
+
+    // Check node id
+    if let Ok(node) = context.lookup(&Identifier(Refer::none("a".into())).into()) {
+        let id = node.id();
+        assert_eq!(id.id(), "a");
+    }
+
+    // Get part definition for symbol `a`
+    let definition = match &node.borrow().def {
+        SymbolDefinition::Part(definition) => definition.clone(),
+        _ => panic!("Symbol is not a part"),
+    };
+
+    // Call `a` with `b = 3.0`
+    let nodes = definition
+        .call(
+            &call_argument_value_list("b = 3.0", &mut context),
+            &mut context,
+        )
+        .expect("Valid nodes");
+
+    assert_eq!(nodes.len(), 1, "There should be one node");
+
+    fn check_node_property_b(node: &ModelNode, value: f64) {
+        if let Element::Object(ref object) = *node.borrow().element() {
+            assert_eq!(
+                object
+                    .get_property_value(&Identifier(Refer::none("b".into())))
+                    .expect("Property `b`"),
+                &Value::Scalar(value)
+            );
+        } else {
+            panic!("Object node expected")
+        }
+    }
+
+    // Test if resulting object node has property `b` with value `3.0`
+    check_node_property_b(nodes.first().expect("Node expected"), 3.0);
+
+    // Call `a` with `b = [1.0, 2.0]` (multiplicity)
+    let nodes = definition
+        .call(
+            &call_argument_value_list("b = [1.0, 2.0]", &mut context),
+            &mut context,
+        )
+        .expect("Valid nodes");
+
+    assert_eq!(nodes.len(), 2, "There should be two nodes");
+
+    check_node_property_b(nodes.first().expect("Node expected"), 1.0);
+    check_node_property_b(nodes.get(1).expect("Node expected"), 2.0);
+}
+
+#[test]
+fn part_explicit_init_call() {
+    use microcad_lang::diag::Diag;
+    use microcad_lang::*;
+
+    let mut context = crate::context_for_file("syntax/part/explicit_init.µcad");
+    let node = context
+        .lookup(&qualified_name("circle"))
+        .expect("Node expected");
+
+    // Get part definition for symbol `a`
+    let definition = match &node.borrow().def {
+        SymbolDefinition::Part(definition) => definition.clone(),
+        _ => panic!("Symbol is not a part"),
+    };
+
+    // Helper function to check if the object node contains a property radius with specified value
+    fn check_node_property_radius(node: &model_tree::ModelNode, value: f64) {
+        if let model_tree::Element::Object(ref object) = *node.borrow().element() {
+            log::trace!("Object: {object}");
+            assert_eq!(
+                object
+                    .get_property_value(&Identifier::no_ref("radius"))
+                    .expect("Property `radius`"),
+                &value::Value::Scalar(value)
+            );
+        } else {
+            panic!("Object node expected")
+        }
+    }
+
+    // Call `circle(radius = 3.0)`
+    {
+        let nodes = definition
+            .call(
+                &call_argument_value_list("radius = 3.0", &mut context),
+                &mut context,
+            )
+            .expect("A valid value");
+        assert_eq!(nodes.len(), 1, "There should be one node");
+        check_node_property_radius(nodes.first().expect("Node expected"), 3.0);
+    }
+
+    // Call `circle(r = 3.0)`
+    {
+        let nodes = definition
+            .call(
+                &call_argument_value_list("r = 3.0", &mut context),
+                &mut context,
+            )
+            .expect("Valid nodes");
+        assert_eq!(nodes.len(), 1, "There should be one node");
+        check_node_property_radius(nodes.first().expect("Node expected"), 3.0);
+    }
+
+    // Call circle(d = 6.0)`
+    {
+        let nodes = definition
+            .call(
+                &call_argument_value_list("d = 6.0", &mut context),
+                &mut context,
+            )
+            .expect("Valid nodes");
+        assert_eq!(nodes.len(), 1, "There should be one node");
+        check_node_property_radius(nodes.first().expect("Node expected"), 3.0);
+    }
+
+    // Call `circle(d = [1.0, 2.0])` (multiplicity)
+    {
+        let nodes = definition
+            .call(
+                &call_argument_value_list("d = [1.0, 2.0]", &mut context),
+                &mut context,
+            )
+            .expect("Valid nodes");
+        assert_eq!(nodes.len(), 2, "There should be two nodes");
+        check_node_property_radius(nodes.first().expect("Node expected"), 0.5);
+        check_node_property_radius(nodes.get(1).expect("Node expected"), 1.0);
+    }
+
+    // Call `circle()` (missing arguments)
+    {
+        let nodes = definition
+            .call(&CallArgumentValueList::default(), &mut context)
+            .expect("Valid nodes");
+        assert_eq!(nodes.len(), 0, "There should no nodes");
+        log::trace!("{}", context.diagnosis());
+    }
+}
